Split “util” packages into something more meaningful
