+++ src/main/java/com/instana/operator/client/KubernetesClientProducer.java
@@ -38,7 +38,7 @@ import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
 import static com.instana.operator.env.NamespaceProducer.OPERATOR_NAMESPACE;
-import static com.instana.operator.util.StringUtils.isEmpty;
+import static com.instana.operator.string.Contents.isEmpty;
 import static okhttp3.ConnectionSpec.CLEARTEXT;
 
 @ApplicationScoped
