+++ src/test/java/com/instana/operator/file/ClasspathFile.java
@@ -1,4 +1,4 @@
-package com.instana.operator.util;
+package com.instana.operator.file;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,9 +6,9 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.stream.Collectors;
 
-public class FileUtil {
-  public static String readFromClasspath(String resource) throws IOException {
-    try (InputStream input = FileUtil.class.getResourceAsStream(resource)) {
+public class ClasspathFile {
+  public static String read(String resource) throws IOException {
+    try (InputStream input = ClasspathFile.class.getResourceAsStream(resource)) {
       try (BufferedReader buffer = new BufferedReader(new InputStreamReader(input))) {
         return buffer.lines().collect(Collectors.joining("\n"));
       }
